Split conn_manager_impl_test to avoid gcc build timeout
